+++ example/android/app/src/main/kotlin/com/example/animated_clipper_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.animated_clipper_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}