+++ lib/src/path_builders/slide/slide.dart
@@ -1,0 +1,6 @@
+// TODO - Add PathBuilderSlideAngle
+
+export 'slide_down.dart';
+export 'slide_left.dart';
+export 'slide_right.dart';
+export 'slide_up.dart';